Tag subnets.  Create network load balancers with listeners and security groups for the API. Add DNS configuration
Private and public subnets a tagged with:

kubernetes.io/cluster/<cluster-name>-<random string> = "shared"

The random string is the same for all subnets, and the rest of components tagged in the future.
The value of the tag is "shared" because this is an UPI installation, I guess.

Load balancers of type network have been created for the internal and external API entry points.  Listeners and target groups have also been created, although no targets are yet added.

IAM roles and policies were created but later commented out because they are not required by Terraform to add the previously mentioned tags

A public DNS zone is created, and A record for the API entry point is added in it.
A private DNS zone is created, and A records for api and api-int are added in it.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>