Create EC2 instances for Bootstrap and masters.  Register the instances with the NLBs target groups
A network interfaces is created for each instance,  it is associated with a subnet and the appropriate security groups for the instance role.

The type of instance is defined in the code, A variable should be used instead

A map variable was created to get the RHCOS AMI corresponding to the AWS region.  This variable is static, does not update if the AMIs change.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>